MainMenu.xib: Don't resize the window when path is long
Previously, when the file path of the folder being served was long, it
would resize the window to fit the path.

Prevent resizing, and truncate the path label so that it truncates the
beginning of the path and keeps the end.

Thanks to Stephan (https://stackoverflow.com/users/565336/stephan) on
Stack Overflow for the tip about setting

> Content Compression Resistance Priority of the NSTextView less than
> NSLayoutPriorityWindowSizeStayPut (e.g. 499). Otherwise your
> NSTextView will not truncate its content.

(https://stackoverflow.com/questions/4171465/truncate-nstextview/13977990#13977990)